refactor: support multiple msg formats
