+++ app/index.js
@@ -2,7 +2,7 @@ const server = require('./server')
 const { log, logError } = require('./services/logger')
 
 const init = async () => {
-  const agreementChangedAction = require('./messaging/agreement-changed')
+  const agreementChangedAction = require('./messaging/determine-agreement-version')
   require('./messaging/receivers').startAgreementChanged(agreementChangedAction)
 
   await server.start()
