+++ app/services/payment-calculator.js
@@ -34,38 +34,12 @@ function selectExpression (ctx) {
  * `expression`- an expression that will be evaluated by `expression-eval`, if
  * the condition is true
  */
-function calcPayment (ctx) {
+function calculatePayment (ctx) {
   const exp = selectExpression(ctx)
   log(`Calculation expression used for '${ctx.id}' - '${exp}'`)
   return evalExpression(ctx, exp)
 }
 
-module.exports = async function (msg) {
-  const { body, correlationId } = msg
-
-  const { standards } = body
-  Object.entries(standards).forEach(([id, standard]) => {
-    const payment = calcPayment(standard)
-    if (!isNaN(payment)) {
-      standards[id].payment = payment
-    }
-    standards[id].optionalActions.forEach(a => {
-      const payment = calcPayment(a)
-      if (!isNaN(payment)) {
-        a.payment = payment
-      }
-    })
-  })
-
-  const selectedStandards = Object.values(standards).filter(s => s.selected)
-  const standardsPayment = selectedStandards.reduce((acc, cur) => { acc += cur.payment; return acc }, 0)
-  const actionsPayment = selectedStandards.reduce((acc, cur) => { acc += cur.optionalActions.reduce((aacc, acur) => { aacc += acur.payment; return aacc }, 0); return acc }, 0)
-
-  body.payments = {
-    actionsPayment,
-    standardsPayment,
-    totalPayment: actionsPayment + standardsPayment
-  }
-  console.log(JSON.stringify(body, null, 2))
-  await require('./senders').updateAgreement({ body, correlationId })
+module.exports = {
+  calculatePayment
 }
